ksupport: fix rpc, cache signature (FFI)
The reason of the borrow stuff is explained in https://git.m-labs.hk/M-Labs/artiq-zynq/issues/76 (artiq-zyna repo).
As for `cache_get()`, compiler will perform stack allocation to pre-allocate the returned structure, and pass to cache_get alongside the `key`.
However, ksupport fails to recognize the passed memory, so it will always assume the passed memory as the key.