test: remove notice for test_array_i32/i64 in RPCHostRoundtripTypeTest
Notice: It's messy and hard to understand at fist, but for test_i32_bad and test_i64_bad, the constant themselves are valid if seeing them as their unsigned counterpart, but will be promoted if seen as a signed integer, e.g. 2147483648 is actually promoted as i64, hence why they are seen as bad.

Signed-off-by: Steve Fan <sf@m-labs.hk>