dashboard: handle more situations on default number resolution
If we have both min and max, and 0 falls in the number line, then set default value to 0
Then if we have a min but not a max, set default value to min if it is positive.
So if we instead have a max but not a min, but we have a negative max, set default value to max.
Otherwise, we can go to 0 instead.

Signed-off-by: Steve Fan <sf@m-labs.hk>