rpc_proto: backport of #160 from artiq-zynq
Explanation of why using isize::rem_euclid: The formula of padding is actually (-size) % alignment, but the modulo is actually the remainder, calculated altogether as the byproduct of integer division. So it is a little tricky when it comes to negative numbers.

Example: (-5) % 3 could be both -2 and 1, because -5 - (-2) == -3 and -5 - 1 == -6. The nature of integer division favors -2 but we wanted 1.

So we have to use euclidean division that always emit positive modulo.

A little question: why do we need to add 4 for Tag::Int64 and Tag::Float64?

Signed-off-by: Steve Fan <sf@m-labs.hk>