comm_moninj: set lifetime for socket
For a TCP socket would not disconnect immediately if the connection to core device is suddenly disrupted (e.g. power loss, panic), because we have TCP retransmission (RTT) and SYN retry window as some form of compromises in difficult network situations (and this explains why the stream drops only after 10 seconds or so).

As such it is better to set SO_KEEPALIVE in order to stop the proxy as soon as the connection stops.

Signed-off-by: Steve Fan <sf@m-labs.hk>