proxy_moninj: change PingTarget to HealthCheckService
The old PingTarget becomes a health check service, which is needed for the dashboard overhaul to disable widgets temporarily until reconnection.

It used to be part of the data sync struct before code redesign, but I figured reimplementing it in RPC would come in handy. This could also make good use of the existing RPC tool to quickly inspect what component of the proxy is not in good shape.

Example use case: `sipyco_rpctool localhost 2384 call healthy`

Signed-off-by: Steve Fan <sf@m-labs.hk>