urukul: dont activate state on read operation
The state is invalid if you read the address on the filter and do a write next. So we need to avoid entering read. By looking up the data sheet of AD9910/AD9912 we know that the read mode bit is MSB. However, for SPI they are shifted towards a 32-bit size, that's why it is apparently not working...

Signed-off-by: Steve Fan <sf@m-labs.hk>