SUServo: pipelined coherent phase tracking mode.
Extension of the per-channel control interface by a flag en_pt, serving
to enable coherent DDS phase mode. This works by tracking the
DDS-internal phase accumulator as well as the servo runtime to calculate
the POW accordingly. The reference time for this calculation can be set
for each output channel individually in real time.