Fix throwing non-embedded builtin exceptions
comm_kernel currently encodes all builtin exceptions with an id of 0.
However, lookup for these exceptions happens from the
artiq.coredevice.exceptions module.

This means if you try to throw a builtin exception not in the embedding
map, then it crashes when receiving it on the host. This also prevents
you catching these exceptions in kernel code.

We now encode only use an exn id of 0 if the exception is in the
exceptions module (and thus already in the embedding map).

Signed-off-by: Jonathan Coates <jonathan.coates@oxionics.com>