Avoid quadratic behaviour in `_unify_attribute` (#2673)
ARTIQ maintains a list of all known instances of a particular class, and
after each attribute access, `_unify_attribute` is called to check all
these instances have this attribute (and it's the right type).

While the attribute type computation is cached, this is still O(nm)
(with n being the number of instances, and m the number of attribute
lookups). For something like ndscan's `FloatParamHandle.get`, you can
have a lot of both.

This commit changes `_unify_attribute` to only check the attributes of
new instances, effectively lowering the complexity to O(n). This
provides a small (5%) boost to compile times.

Signed-off-by: Jonathan Coates <jonathan.coates@oxionics.com>