gui/fuzzy_select: QMenu.{show->popup}() to fix positioning on Windows 11
Even though the menu had initially been shown using popup() already,
the hide()/show() pair caused the menu to appear at (0, 0) if on the
primary screen, and not at all if on another screen. This was on
Qt 6.9 and Windows 11.

The Qt docs do warn against using show(), but it was not clear whether
this would apply to just temporarily hiding the menu, and the previous
implementation worked fine on Qt 5/Windows 10.