Add support for static methods for artiq
Pass through the python builtin `staticmethod` to nac3. The decoration
does not determine the hardware the function will be ran on, so to
specify this use the `@rpc`/`@portable`/`@kernel` flags as usual. If not
decorated with anything it will, as any other function, be only visible
to the host.