adf5356: nac3ify calculate_pll() properly
Allow f_vco to be a float while keeping f_pfd as an int64, at least
until this changes in mainline ARTIQ. This requires casting f_pfd to a
float to enable e.g. the modulo operation with f_vco, as NAC3 doesn't
allow a modulo operation between a float and an int64.

Note: modulo operations with floats are analogous to their integer
counterparts; for floats a and b, q * b + a % b is very close to a,
where q = a // b.