adf5356: Recast f_pfd to int64 in calculate_pll
Under the @portable context, a float could be passed in to the f_pfd
parameter even when type-annotated as an int64. We want to be able to
specify f_pfd as a literal like 125*MHz, which is a float (due to the Hz
unit) even if it is an integral value, hence the casting.